Restructure code-base
